third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-4/styles-update-links-4-expected.txt

Tests that links are updated property when editing pseudo element property.


Running: selectNodeWithPseudoElements

Running: testPseudoSectionPropertyEdit


#### AFTER PROPERTY INSERTED ####


[expanded] 
element.style { ()

[expanded] 
article, aside, footer, header, hgroup, main, nav, section { (user agent stylesheet)
    display: block;

======== Pseudo ::before element ========
[expanded] 
#pseudo::before { (<style>…</style>)
    color: blue;

======== Pseudo ::after element ========
[expanded] 
#pseudo::after { (<style>…</style>)
    border: 1px solid black;

[expanded] 
#pseudo::after { (<style>…</style>)
/-- overloaded --/     pseudo-property: "12";
/-- overloaded --/     PROPERTY: INSERTED;
    color: red;

0 rule ranges are equal.
3 rule ranges are equal.
3 rule ranges are equal.
5 rule ranges are equal.
